+++ modules/system/metadata.sh
@@ -1,0 +1,3 @@
+guardian_module_id="system"
+guardian_module_version="1"
+guardian_module_description="Basic Linux host checks"