+++ modules/systemd/metadata.sh
@@ -1,0 +1,3 @@
+guardian_module_id="systemd"
+guardian_module_version="1"
+guardian_module_description="systemd health checks"